requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
groq>=0.4.0
litellm>=1.0.0
crewai>=0.74.0
crewai-tools>=0.12.0
tavily-python>=0.5.0
python-dotenv>=1.0.1
requests>=2.32.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
rich>=13.7.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0


